Add support for Rails 4 & 5
There are subtle differences between Rails 3 and 4 in the way bind
values are handled.  A small fix supports both versions.  Rails 5
completely chnages the internals this gem relies on.  However, Rails 5
comes with an #or method on relations that mostly makes this gem
superfluous.  Nevertheless, we'll continue to support Rails 5 to ease
upgrades.