Ensure where_any_of() returns no records
It was discovered that passing no arguments to `where_any_of` ends up
adding no constraints to the current scope, but this is not intuitive
and is not consistent with Ruby's Enumerable#any? which requires at
least one condition to be true.

This fix ensures that passing no arguments to `where_any_of` results in
no records being returned.